+++ content-management-system/marketplace-page/order-mgmt/delete-orders.php
@@ -1,0 +1,21 @@
+<?php
+require '../admin-required.php';
+require '../../parts/connect_db.php';
+
+$id = isset($_GET['id']) ? intval($_GET['id']) : 0;
+
+if (empty($id)) {
+    header('Location: order-list.php');
+    exit;
+}
+
+$sql = "DELETE FROM `orders` WHERE id=$id";
+$pdo->query($sql);
+
+if (empty($_SERVER['HTTP_REFERER'])) {
+    $come_from = 'order-list.php';
+} else {
+    $come_from = $_SERVER['HTTP_REFERER']; // 從哪裡來, 回哪裡去
+}
+
+header("Location: $come_from");
