+++ content-management-system/parts/connect_db.php
@@ -16,3 +16,7 @@ $pdo_options = [
 ];
 
 $pdo = new PDO($dsn, $db_user, $db_pass, $pdo_options);
+
+if (!isset($_SESSION)) {
+    session_start();
+}
